Move runUpgrades.
